utils and custom
[x] create uitils files to make main workflow neat
[x] create custom regressors paths
